moved packet handling to its own module
